+++ src/commands/configuration.ts
@@ -3,23 +3,23 @@ import { ResponseMessage, NamedMessage } from '../message'
 import { AbstractCommand } from './abstractCommand'
 
 export interface ConfigurationCommandResponse {
-    videoInput: string
-    audioInput: string
-    fileFormat: string
+	videoInput: string
+	audioInput: string
+	fileFormat: string
 }
 
 export class ConfigurationCommand extends AbstractCommand {
-    expectedResponseCode = SynchronousCode.DeviceInfo
+	expectedResponseCode = SynchronousCode.DeviceInfo
 
-    videoInput?: string
-    audioInput?: string
-    fileFormat?: string
+	videoInput?: string
+	audioInput?: string
+	fileFormat?: string
 
 	deserialize (msg: ResponseMessage) {
 		const res: ConfigurationCommandResponse = {
 			videoInput: msg.params['video input'],
-            audioInput: msg.params['audio input'],
-            fileFormat: msg.params['file format']
+			audioInput: msg.params['audio input'],
+			fileFormat: msg.params['file format']
 		}
 		return res
 	}
@@ -27,20 +27,20 @@ export class ConfigurationCommand extends AbstractCommand {
 	constructor (videoInput?: string, audioInput?: string, fileFormat?: string) {
 		super()
 
-        this.videoInput = videoInput
-        this.audioInput = audioInput
-        this.fileFormat = fileFormat
+		this.videoInput = videoInput
+		this.audioInput = audioInput
+		this.fileFormat = fileFormat
 	}
 
 	serialize () {
 		const res: NamedMessage = {
 			name: 'jog',
 			params: {}
-        }
+		}
 
-        if (this.videoInput) res.params.videoInput = this.videoInput
-        if (this.audioInput) res.params.audioInput = this.audioInput
-        if (this.fileFormat) res.params.fileFormat = this.fileFormat
+		if (this.videoInput) res.params.videoInput = this.videoInput
+		if (this.audioInput) res.params.audioInput = this.audioInput
+		if (this.fileFormat) res.params.fileFormat = this.fileFormat
 
 		return res
 	}
