+++ src/commands/index.ts
@@ -7,7 +7,7 @@ export * from './format'
 export * from './goTo'
 export * from './jog'
 export * from './notify'
-export * from './playCommand'
+export * from './play'
 export * from './record'
 export * from './slotInfo'
 export * from './slotSelect'
