+++ src/commands/jog.ts
@@ -2,23 +2,23 @@ import { NamedMessage } from '../message'
 import { AbstractCommandNoResponse } from './abstractCommand'
 
 export class JogCommand extends AbstractCommandNoResponse {
-    timecode?: string
+	timecode?: string
 
 	constructor (timecode?: string) {
 		super()
 
-        this.timecode = timecode
+		this.timecode = timecode
 	}
 
 	serialize () {
 		const res: NamedMessage = {
 			name: 'jog',
 			params: {}
-        }
+		}
 
-        if (this.timecode === undefined) throw new Error('JogCommand needs timecode')
+		if (this.timecode === undefined) throw new Error('JogCommand needs timecode')
 
-        res.params.timecode = this.timecode
+		res.params.timecode = this.timecode
 
 		return res
 	}
